Bring back the clean-up done previously for download
